Change BPList design
