+++ frontend/src/app/business-plans/page.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Metadata } from 'next'
 
-import BPList from '@ors/components/manage/Blocks/BusinessPlans/BPList'
+import BPList from '@ors/components/manage/Blocks/BusinessPlans/BPList/BPList'
 import HeaderTitle from '@ors/components/theme/Header/HeaderTitle'
 import PageWrapper from '@ors/components/theme/PageWrapper/PageWrapper'
 import { PageHeading } from '@ors/components/ui/Heading/Heading'
@@ -13,7 +13,7 @@ export const metadata: Metadata = {
 
 export default async function BusinessPlans() {
   return (
-    <PageWrapper className="max-w-screen-xl lg:px-0">
+    <PageWrapper className="max-w-screen-xl xl:px-0">
       <HeaderTitle>
         <PageHeading className="mx-auto max-w-screen-xl">
           Business Plans
